requirements.txt
python-2.7.18rc1
Django==2.2.10
django-heroku==0.3.0
djangorestframework==3.11.0
gunicorn==20.0.4
psycopg2==2.8.5
uWSGI==2.0.18
whitenoise==5.1.0